--- src/components/AmountInput/styles.js
@@ -1,4 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-});
