--- src/components/AppText/index.js
@@ -1,2 +1,0 @@
-import AppText from './AppText';
-export default AppText;
