--- src/components/BatchList/index.js
@@ -1,2 +1,0 @@
-import BatchList from './BatchList';
-export default BatchList;
