--- src/components/BatchSummary/index.js
@@ -1,2 +1,0 @@
-import BatchSummary from './BatchSummary';
-export default BatchSummary;
