--- src/components/COINiDTransport/index.js
@@ -1,3 +1,0 @@
-import COINiDTransport from './COINiDTransport';
-
-export default COINiDTransport;
