--- src/components/CheckBoxSelect/index.js
@@ -1,2 +1,0 @@
-import CheckBoxSelect from './CheckBoxSelect';
-export default CheckBoxSelect;
