--- src/components/ConnectionStatus/index.js
@@ -1,2 +1,0 @@
-import ConnectionStatus from './ConnectionStatus';
-export default ConnectionStatus;
