--- src/components/DialogTitle/index.js
@@ -1,2 +1,0 @@
-import DialogTitle from './DialogTitle';
-export default DialogTitle;
