+++ src/components/FeeSlider.js
@@ -1,13 +1,46 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { View, Animated, Easing } from 'react-native';
+import {
+  StyleSheet, View, Animated, Easing,
+} from 'react-native';
 import Big from 'big.js';
 import { Slider } from 'react-native-elements';
-import { Text, RowInfo } from '..';
-import { numFormat } from '../../utils/numFormat';
-import FeeHelper from '../../utils/feeHelper';
-import styles from './styles';
-import { colors } from '../../config/styling';
+import { Text, RowInfo } from '.';
+import { numFormat } from '../utils/numFormat';
+import FeeHelper from '../utils/feeHelper';
+import { colors, fontSize } from '../config/styling';
+
+const styles = StyleSheet.create({
+  estimationText: {
+    fontSize: fontSize.small,
+    color: colors.getTheme('light').fadedText,
+  },
+  warningText: {
+    fontSize: fontSize.small,
+    color: colors.getTheme('light').warning,
+  },
+  satByteDot: {
+    backgroundColor: colors.getTheme('light').button,
+    borderRadius: 18,
+    position: 'absolute',
+    zIndex: 100,
+    paddingTop: 6,
+    paddingBottom: 7,
+    paddingHorizontal: 10,
+  },
+  satByteText: {
+    color: colors.getTheme('light').buttonText,
+  },
+  thumb: {
+    width: 24,
+    height: 24,
+    borderRadius: 12,
+  },
+  track: {
+    backgroundColor: colors.lightGray,
+    height: 2,
+  },
+});
 
 export default class FeeSlider extends PureComponent {
   constructor(props, context) {
