--- src/components/FeeSlider/index.js
@@ -1,2 +1,0 @@
-import FeeSlider from './FeeSlider';
-export default FeeSlider;
