--- src/components/FontScale/index.js
@@ -1,2 +1,0 @@
-import FontScale from './FontScale';
-export default FontScale;
