--- src/components/Graph/index.js
@@ -1,2 +1,0 @@
-import Graph from './Graph';
-export default Graph;
